refactor authentication flow by replacing PasswordForm with PasswordCheck component for improved password handling
